+++ src/app/(app)/dashboard/[accountSlug]/settings/members/_components/UserPlaceholder.tsx
@@ -1,0 +1,14 @@
+const UserPlaceholder = () => (
+  <div className="flex items-center justify-between space-x-3 px-4 py-3 sm:px-8">
+    <div className="flex items-center space-x-3">
+      <div className="h-10 w-10 animate-pulse rounded-full bg-gray-200" />
+      <div className="flex flex-col">
+        <div className="h-4 w-24 animate-pulse rounded bg-gray-200" />
+        <div className="mt-1 h-3 w-32 animate-pulse rounded bg-gray-200" />
+      </div>
+    </div>
+    <div className="h-3 w-24 animate-pulse rounded bg-gray-200" />
+  </div>
+);
+
+export default UserPlaceholder;
